fix bug: do not terminate an already terminated process

+++ babyai/rl/utils/penv.py
@@ -56,4 +56,5 @@ class ParallelEnv(gym.Env):
 
     def __del__(self):
         for p in self.processes:
-            p.terminate()
+            if p.is_alive():
+                p.terminate()
